Runs, does not work.

+++ src/entity.rs
@@ -1,2 +1,2 @@
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct Entity(pub u32);
+pub struct Entity(pub usize);
